hexa a gerer avec les negatifs

+++ ft_printf.c
@@ -6,7 +6,7 @@
 /*   By: norabino <norabino@student.42perpignan.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/13 11:42:04 by norabino          #+#    #+#             */
-/*   Updated: 2024/11/15 15:57:20 by norabino         ###   ########.fr       */
+/*   Updated: 2024/11/15 15:59:42 by norabino         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -49,9 +49,9 @@ void	ft_conversions(va_list ap, const char *format, int	*i, int *len)
 	if (format[*i + 1] && format[*i + 1] == 'd')
 		*len += ft_print_d((long)va_arg(ap, long));
 	if (format[*i + 1] && format[*i + 1] == 'x')
-		*len += ft_print_x((unsigned long)va_arg(ap, unsigned long));
+		*len += ft_print_x((unsigned long long)va_arg(ap, unsigned long long));
 	if (format[*i + 1] && format[*i + 1] == 'X')
-		*len += ft_print_X((unsigned long)va_arg(ap, unsigned long));
+		*len += ft_print_X((unsigned long long)va_arg(ap, unsigned long long));
 }
 
 int	ft_printf(const char *format, ...)
